+++ gdsfactory/simulation/simphony/components/bend_euler.py
@@ -9,7 +9,7 @@ def bend_euler(
     radius: float = 10.0,
     width: float = 0.5,
     thickness: float = 0.22,
-    angle: int = 90,
+    angle: float = 90,
     sw_angle: float = 90.0,
     **kwargs,
 ):
